Disable signature force delete and add signature form to profile

+++ app/Filament/Superuser/Resources/SignatureResource/Pages/EditSignature.php
@@ -47,7 +47,7 @@ class EditSignature extends EditRecord
                     mkdir(storage_path('app/signatures/certificates'), recursive: true);
                 }
 
-                if (file_exists(storage_path('app/'.$this->record->certificate))) {
+                if ($this->record->certificate && file_exists(storage_path('app/'.$this->record->certificate))) {
                     unlink(storage_path('app/'.$this->record->certificate));
                 }
